requirements.txt
pygame==2.1.0
openai==0.28.1
requests==2.31.0
Pillow==9.5.0
